requirements.txt
streamlit
tensorflow
numpy
opencv-python
pillow
requests
streamlit-lottie
openai
